+++ lib/models/asyncapi.js
@@ -1,4 +1,4 @@
-const { createMapOfType, getMapValueOfType, mix } = require('../models-utils');
+const { createMapOfType, getMapValueOfType, mix } = require('./utils');
 
 const Base = require('./base');
 const Info = require('./info');
