+++ lib/models/message-traitable.js
@@ -1,4 +1,4 @@
-const { getMapValueOfType, mix } = require('../models-utils');
+const { getMapValueOfType, mix } = require('./utils');
 
 const Base = require('./base');
 const Schema = require('./schema');
