+++ lib/models/security-scheme.js
@@ -1,4 +1,4 @@
-const { createMapOfType, mix } = require('../models-utils');
+const { createMapOfType, mix } = require('./utils');
 
 const Base = require('./base');
 const OAuthFlow = require('./oauth-flow');
