+++ lib/models/server-variable.js
@@ -1,4 +1,4 @@
-const { mix } = require('../models-utils');
+const { mix } = require('./utils');
 
 const Base = require('./base');
